requirements.txt
python-dotenv
yfinance
supabase
pytz
